+++ src/main/java/nl/gjorgdy/interactive_world/InteractiveWorld.java
@@ -37,6 +37,7 @@ public class InteractiveWorld implements ModInitializer {
 	public static boolean glowBerriesGlowEffect = true;
 	public static boolean blastProofTintedGlass = true;
 	public static boolean toggleItemFrameVisibility = true;
+	public static boolean clickThroughWallSign = true;
 
 	public static int fallDistanceTrampleFarmland = 8;
 	public static int glowBerryEffectTimeTicks = 160;