+++ changelog.md
@@ -1,4 +1,6 @@
 ﻿A Sulfur World
 ---
 
+- Clicking on an invisible item frame while not crouching interacts with the block behind it.
+- Clicking on a wall sign on an interactable block will ignore the sign if not crouching.
 - Updated to 26.2